fix: 🐛 Handling Unsigned Value in DivU

+++ src/execution/stack.rs
@@ -372,8 +372,8 @@ impl LabelStack{
                             None
                         },
                         Instr::I32DivU => {
-                            let rhs = self.value_stack.pop().unwrap().to_i32();
-                            let lhs = self.value_stack.pop().unwrap().to_i32();
+                            let rhs = self.value_stack.pop().unwrap().to_i32() as u32;
+                            let lhs = self.value_stack.pop().unwrap().to_i32() as u32;
                             self.value_stack.push(
                                 Val::Num(Num::I32(lhs.checked_div(rhs).ok_or_else(|| RuntimeError::ZeroDivideError)? as i32))
                             );
